Add files via upload
PicoCTF 2018:
Forensics Warmup Challenge - 1
=> Can you unzip the file for me and retrieve the flag?

Hints: Make sure you submit the flag as picoCTF{XXXXX}

Solution: 

Let's try and unzip the file to know what lies inside. 
To unzip the file:

a) go to the location of the file where it is downloaded and righ click on it to find the 'unzip' option. Click on it to unzip.

b) open terminal and type in the following commands - 
	(i) unzip flag.zip 	//to unzip and extract contents of the file
	(ii) eog flag.jpg 	// if the flag is of .jpg format. You will come to know about the image type once you extract the files.